feat: add control in order to avoid deletion of not cancelled orders

+++ business-app/back/src/main/java/com/adeo/demo/order/backend/services/OrderService.java
@@ -74,6 +74,13 @@ public class OrderService {
 
     @Transactional
     public void deleteOrder(Long orderId) {
+        Optional<Order> existingOrder = orderRepository.findById(orderId); // Ensure the order exists before deletion
+        if (existingOrder.isEmpty() ){
+            throw new IllegalArgumentException("Order not found with ID: " + orderId);
+        }
+        if(existingOrder.get().getStatus() != OrderStatus.FINISHED || existingOrder.get().getStatus() != OrderStatus.CANCELLED) {
+            throw new IllegalStateException("Cannot delete order that is not FINISHED or CANCELLED");
+        }
         orderRepository.deleteById(orderId);
     }
